Adjust YaST namespaces
We had different spellings for YaST (vs Yast). I fixed that.
Additionally, I moved a namespaced class into the proper folder.

Looking at the current code, I'm not quite sure whether we need the
YaST namespace at all. IMHO we should get rid of the Extension class
anyways (we don't have that in SCC - it's a Product). If we do that,
then there is no need for a namespace.

Nevertheless, for the moment this seems to be more consistent.