Prevent NoMethodError when running SUSEConnect on an s390x LPAR
When running directly on an LPAR the read_values command doesn't list
VM00 Control Program. SUSEConnect was expecting this to be set in any
case and thus failed with NoMethodError.
With this commit such a case will be covered by setting the hypervisor
to nil.

Error message on a SLE12 system:

  s390lp2:~ # SUSEConnect --regcode "xxx"
  SUSEConnect error: NoMethodError: undefined method `strip' for nil:NilClass