Make base product version zypp lock independent
Looking up the regcode in the step definitions was using the gem itself
(which is a smell) and we just got bitten by this during the prophet
tests, because the gem calls zypper to know the base product [version]
and this cannot work during the zypp locked test.

I already pass in the product version as an environment variable for
integration testing so I am now using this in the step definitions and
keeping zypper calls for a fallback.

Passing this env var in to prophet prevents the failing call to zypper.