Correctly set the proxy user & pass in Ruby 2.5
Ruby 2.5 changes the way proxies behave from Ruby 2.1.
In Ruby 2.1, Net::HTTP would get the proxy URL from the http_proxy env
var, and then one could set proxy_user and proxy_password directly on
the Net::HTTP object.
In Ruby 2.5, once the proxy URL comes from the http_proxy env var,
setting proxy_user and proxy_password on the Net::HTTP object does
nothing. It assumes that the proxy URL will have the user and password
as part of the URI (eg. http://user:pass@myproxy).
In SLES, the recommended way of setting the proxy user and password is
through the .curlrc file.
Now, what we do is that if we detect that a proxy is being used, we will
re-create the Net::HTTP object, passing all proxy settings to the
initializer.