Change startup_wait_time config
Basically, Aruba will `sleep` the amount of time that is configured in
startup_wait_time on EVERY command run, so setting it to a high number
will make tests terribly slow.
But for some reason, we still need to have set it to SOME number in
order for the announcers defined by
activate_announcer_on_command_failure to work. Otherwise, those
announcers show absolutely no output, even though the command did have
SOME output.
I added a new @slow_startup label because the rollback takes a long time
to run and we need to give it some more time.