Using "successfully run" was not waiting until commands really finished
So it was giving strange errors about "expected (command) to have
finished in time", without giving commands the opportunity to finish. I
could not find a source for this problem, so I just decided to drop it
and go back to "I run". The activate_announcer_on_command_failure option
is only used when using "I successfully run" so it does not make sense
to keep it.