Allow installing openssl 1.0 without removing SUSEConnect
Due to the way openssl was recommended, one could not install openssl
1.0 from the Legacy module without uninstalling SUSEConnect. To bypass
this, they added a resolvable that represents the openssl cli (which is
what SUSEConnect needs).
This resolvable has only been released on SLES 12 SP2+SP3 and SLES 15,
so we need to fall back to the previous recommend on SLES 12 SP0 + SP1.
See: https://bugzilla.suse.com/show_bug.cgi?id=1101470

This change was initially submitted to our Build Service project (see
https://build.opensuse.org/request/show/637066) but it did not account
for the SLE 12 SP0 and SP1 case.