Write Config Even During Failure Conditions
We noticed a problem in RMT that came from a failure logic in
SUSEConnect.

* A server is registered against RMT with SUSEConnect, but that product
  isn't mirrored yet in RMT.
* As the Connect API announce endpoint does not verify that a product is
  mirrored, RMT has no idea that a subsequent call will try to activate
  an unmirrored product.
* The server is announced to RMT, so RMT gave it credentials which are
  saved on the server.
* During the activation phase, RMT tells the server it can't activate,
  so the registration process in SUSEConenct errors and the config isn't
  written.

The problem is that SUSEConnect already got the credentials from the
RMT server, but because it errored, it didn't save /etc/SUSEConnect to
know where those credentials are valid. Instead, we should save
/etc/SUSEConnect, as the system is valid in RMT, it's activation just
isn't and can be fixed without a deregistration.